+++ discord/commands/core.py
@@ -45,14 +45,21 @@ from typing import (
     Union,
 )
 
-from ..channel import _threaded_guild_channel_factory
+from ..channel import PartialMessageable, _threaded_guild_channel_factory
 from ..enums import Enum as DiscordEnum
-from ..enums import MessageType, SlashCommandOptionType, try_enum
+from ..enums import (
+    IntegrationType,
+    InteractionContextType,
+    MessageType,
+    SlashCommandOptionType,
+    try_enum,
+)
 from ..errors import (
     ApplicationCommandError,
     ApplicationCommandInvokeError,
     CheckFailure,
     ClientException,
+    InvalidArgument,
     ValidationError,
 )
 from ..member import Member
@@ -61,7 +68,7 @@ from ..object import Object
 from ..role import Role
 from ..threads import Thread
 from ..user import User
-from ..utils import MISSING, async_all, find, maybe_coroutine, utcnow
+from ..utils import MISSING, async_all, find, maybe_coroutine, utcnow, warn_deprecated
 from .context import ApplicationContext, AutocompleteContext
 from .options import Option, OptionChoice
 
@@ -226,10 +233,35 @@ class ApplicationCommand(_BaseCommand, Generic[CogT, P, T]):
             "__default_member_permissions__",
             kwargs.get("default_member_permissions", None),
         )
-        self.guild_only: bool | None = getattr(
-            func, "__guild_only__", kwargs.get("guild_only", None)
-        )
         self.nsfw: bool | None = getattr(func, "__nsfw__", kwargs.get("nsfw", None))
+
+        integration_types = getattr(
+            func, "__integration_types__", kwargs.get("integration_types", None)
+        )
+        contexts = getattr(func, "__contexts__", kwargs.get("contexts", None))
+        guild_only = getattr(func, "__guild_only__", kwargs.get("guild_only", MISSING))
+        if guild_only is not MISSING:
+            warn_deprecated(
+                "guild_only",
+                "contexts",
+                "2.6",
+                reference="https://discord.com/developers/docs/change-log#userinstallable-apps-preview",
+            )
+        if contexts and guild_only:
+            raise InvalidArgument(
+                "cannot pass both 'contexts' and 'guild_only' to ApplicationCommand"
+            )
+        if self.guild_ids and (
+            (contexts is not None) or guild_only or integration_types
+        ):
+            raise InvalidArgument(
+                "the 'contexts' and 'integration_types' parameters are not available for guild commands"
+            )
+
+        if guild_only:
+            contexts = {InteractionContextType.guild}
+        self.contexts: set[InteractionContextType] | None = contexts
+        self.integration_types: set[IntegrationType] | None = integration_types
 
     def __repr__(self) -> str:
         return f"<discord.commands.{self.__class__.__name__} name={self.name}>"
@@ -273,6 +305,33 @@ class ApplicationCommand(_BaseCommand, Generic[CogT, P, T]):
         self._callback = function
         unwrap = unwrap_function(function)
         self.module = unwrap.__module__
+
+    @property
+    def guild_only(self) -> bool:
+        warn_deprecated(
+            "guild_only",
+            "contexts",
+            "2.6",
+            reference="https://discord.com/developers/docs/change-log#userinstallable-apps-preview",
+        )
+        return InteractionContextType.guild in self.contexts and len(self.contexts) == 1
+
+    @guild_only.setter
+    def guild_only(self, value: bool) -> None:
+        warn_deprecated(
+            "guild_only",
+            "contexts",
+            "2.6",
+            reference="https://discord.com/developers/docs/change-log#userinstallable-apps-preview",
+        )
+        if value:
+            self.contexts = {InteractionContextType.guild}
+        else:
+            self.contexts = {
+                InteractionContextType.guild,
+                InteractionContextType.bot_dm,
+                InteractionContextType.private_channel,
+            }
 
     def _prepare_cooldowns(self, ctx: ApplicationContext):
         if self._buckets.valid:
@@ -631,6 +690,9 @@ class SlashCommand(ApplicationCommand):
         Returns a string that allows you to mention the slash command.
     guild_only: :class:`bool`
         Whether the command should only be usable inside a guild.
+
+        .. deprecated:: 2.6
+            Use the :attr:`contexts` parameter instead.
     nsfw: :class:`bool`
         Whether the command should be restricted to 18+ channels and users.
         Apps intending to be listed in the App Directory cannot have NSFW commands.
@@ -654,6 +716,12 @@ class SlashCommand(ApplicationCommand):
     description_localizations: Dict[:class:`str`, :class:`str`]
         The description localizations for this command. The values of this should be ``"locale": "description"``.
         See `here <https://discord.com/developers/docs/reference#locales>`_ for a list of valid locales.
+    integration_types: Set[:class:`IntegrationType`]
+        The type of installation this command should be available to. For instance, if set to
+        :attr:`IntegrationType.user_install`, the command will only be available to users with
+        the application installed on their account. Unapplicable for guild commands.
+    contexts: Set[:class:`InteractionContextType`]
+        The location where this command can be used. Cannot be set if this is a guild command.
     """
 
     type = 1
@@ -881,9 +949,6 @@ class SlashCommand(ApplicationCommand):
         if self.is_subcommand:
             as_dict["type"] = SlashCommandOptionType.sub_command.value
 
-        if self.guild_only is not None:
-            as_dict["dm_permission"] = not self.guild_only
-
         if self.nsfw is not None:
             as_dict["nsfw"] = self.nsfw
 
@@ -891,6 +956,10 @@ class SlashCommand(ApplicationCommand):
             as_dict["default_member_permissions"] = (
                 self.default_member_permissions.value
             )
+
+        if not self.guild_ids and not self.is_subcommand:
+            as_dict["integration_types"] = [it.value for it in self.integration_types]
+            as_dict["contexts"] = [ctx.value for ctx in self.contexts]
 
         return as_dict
 
@@ -1100,6 +1169,9 @@ class SlashCommandGroup(ApplicationCommand):
         isn't one.
     guild_only: :class:`bool`
         Whether the command should only be usable inside a guild.
+
+        .. deprecated:: 2.6
+            Use the :attr:`contexts` parameter instead.
     nsfw: :class:`bool`
         Whether the command should be restricted to 18+ channels and users.
         Apps intending to be listed in the App Directory cannot have NSFW commands.
@@ -1118,6 +1190,12 @@ class SlashCommandGroup(ApplicationCommand):
     description_localizations: Dict[:class:`str`, :class:`str`]
         The description localizations for this command. The values of this should be ``"locale": "description"``.
         See `here <https://discord.com/developers/docs/reference#locales>`_ for a list of valid locales.
+    integration_types: Set[:class:`IntegrationType`]
+        The type of installation this command should be available to. For instance, if set to
+        :attr:`IntegrationType.user_install`, the command will only be available to users with
+        the application installed on their account. Unapplicable for guild commands.
+    contexts: Set[:class:`InteractionContextType`]
+        The location where this command can be used. Unapplicable for guild commands.
     """
 
     __initial_commands__: list[SlashCommand | SlashCommandGroup]
@@ -1177,8 +1255,29 @@ class SlashCommandGroup(ApplicationCommand):
         self.default_member_permissions: Permissions | None = kwargs.get(
             "default_member_permissions", None
         )
-        self.guild_only: bool | None = kwargs.get("guild_only", None)
         self.nsfw: bool | None = kwargs.get("nsfw", None)
+
+        integration_types = kwargs.get("integration_types", None)
+        contexts = kwargs.get("contexts", None)
+        guild_only = kwargs.get("guild_only", MISSING)
+        if guild_only is not MISSING:
+            warn_deprecated("guild_only", "contexts", "2.6")
+        if contexts and guild_only:
+            raise InvalidArgument(
+                "cannot pass both 'contexts' and 'guild_only' to ApplicationCommand"
+            )
+        if self.guild_ids and (
+            (contexts is not None) or guild_only or integration_types
+        ):
+            raise InvalidArgument(
+                "the 'contexts' and 'integration_types' parameters are not available for guild commands"
+            )
+
+        # These are set to None and their defaults are then set when added to the bot
+        self.contexts: set[InteractionContextType] | None = contexts
+        if guild_only:
+            self.guild_only: bool | None = guild_only
+        self.integration_types: set[IntegrationType] | None = integration_types
 
         self.name_localizations: dict[str, str] = kwargs.get(
             "name_localizations", MISSING
@@ -1218,6 +1317,23 @@ class SlashCommandGroup(ApplicationCommand):
     def module(self) -> str | None:
         return self.__module__
 
+    @property
+    def guild_only(self) -> bool:
+        warn_deprecated("guild_only", "contexts", "2.6")
+        return InteractionContextType.guild in self.contexts and len(self.contexts) == 1
+
+    @guild_only.setter
+    def guild_only(self, value: bool) -> None:
+        warn_deprecated("guild_only", "contexts", "2.6")
+        if value:
+            self.contexts = {InteractionContextType.guild}
+        else:
+            self.contexts = {
+                InteractionContextType.guild,
+                InteractionContextType.bot_dm,
+                InteractionContextType.private_channel,
+            }
+
     def to_dict(self) -> dict:
         as_dict = {
             "name": self.name,
@@ -1232,9 +1348,6 @@ class SlashCommandGroup(ApplicationCommand):
         if self.parent is not None:
             as_dict["type"] = self.input_type.value
 
-        if self.guild_only is not None:
-            as_dict["dm_permission"] = not self.guild_only
-
         if self.nsfw is not None:
             as_dict["nsfw"] = self.nsfw
 
@@ -1242,6 +1355,10 @@ class SlashCommandGroup(ApplicationCommand):
             as_dict["default_member_permissions"] = (
                 self.default_member_permissions.value
             )
+
+        if not self.guild_ids and self.parent is None:
+            as_dict["integration_types"] = [it.value for it in self.integration_types]
+            as_dict["contexts"] = [ctx.value for ctx in self.contexts]
 
         return as_dict
 
@@ -1476,6 +1593,9 @@ class ContextMenuCommand(ApplicationCommand):
         The ids of the guilds where this command will be registered.
     guild_only: :class:`bool`
         Whether the command should only be usable inside a guild.
+
+        .. deprecated:: 2.6
+            Use the ``contexts`` parameter instead.
     nsfw: :class:`bool`
         Whether the command should be restricted to 18+ channels and users.
         Apps intending to be listed in the App Directory cannot have NSFW commands.
@@ -1496,6 +1616,10 @@ class ContextMenuCommand(ApplicationCommand):
     name_localizations: Dict[:class:`str`, :class:`str`]
         The name localizations for this command. The values of this should be ``"locale": "name"``. See
         `here <https://discord.com/developers/docs/reference#locales>`_ for a list of valid locales.
+    integration_types: Set[:class:`IntegrationType`]
+        The installation contexts where this command is available. Unapplicable for guild commands.
+    contexts: Set[:class:`InteractionContextType`]
+        The interaction contexts where this command is available. Unapplicable for guild commands.
     """
 
     def __new__(cls, *args, **kwargs) -> ContextMenuCommand:
@@ -1575,8 +1699,9 @@ class ContextMenuCommand(ApplicationCommand):
             "type": self.type,
         }
 
-        if self.guild_only is not None:
-            as_dict["dm_permission"] = not self.guild_only
+        if not self.guild_ids:
+            as_dict["integration_types"] = [it.value for it in self.integration_types]
+            as_dict["contexts"] = [ctx.value for ctx in self.contexts]
 
         if self.nsfw is not None:
             as_dict["nsfw"] = self.nsfw
@@ -1729,20 +1854,12 @@ class MessageCommand(ContextMenuCommand):
         for i, v in _data.items():
             v["id"] = int(i)
             message = v
-        channel = ctx.interaction._state.get_channel(int(message["channel_id"]))
-        if channel is None:
-            author_id = int(message["author"]["id"])
-            self_or_system_message: bool = ctx.bot.user.id == author_id or try_enum(
-                MessageType, message["type"]
-            ) not in (
-                MessageType.default,
-                MessageType.reply,
-                MessageType.application_command,
-                MessageType.thread_starter_message,
+        channel = ctx.interaction.channel
+        if channel.id != int(message["channel_id"]):
+            # we got weird stuff going on, make up a channel
+            channel = PartialMessageable(
+                state=ctx.interaction._state, id=int(message["channel_id"])
             )
-            user_id = ctx.author.id if self_or_system_message else author_id
-            data = await ctx.interaction._state.http.start_private_message(user_id)
-            channel = ctx.interaction._state.add_dm_channel(data)
 
         target = Message(state=ctx.interaction._state, channel=channel, data=message)