+++ discord/permissions.py
@@ -634,6 +634,17 @@ class Permissions(BaseFlags):
         """
         return 1 << 49
 
+    @flag_value
+    def use_external_apps(self) -> int:
+        """:class:`bool`: Returns ``True`` if a member's user-installed apps can show public responses.
+        Users will still be able to use user-installed apps, but responses will be ephemeral.
+
+        This only applies to apps that are also not installed to the guild.
+
+        .. versionadded:: 2.6
+        """
+        return 1 << 50
+
 
 PO = TypeVar("PO", bound="PermissionOverwrite")
 
@@ -754,6 +765,7 @@ class PermissionOverwrite:
         send_voice_messages: bool | None
         set_voice_channel_status: bool | None
         send_polls: bool | None
+        use_external_apps: bool | None
 
     def __init__(self, **kwargs: bool | None):
         self._values: dict[str, bool | None] = {}